refactor:organize create task use case, mark task complete use case, repository interfaces, controllers, and presenters into dedicated packages
