--- src/main/java/interface_adapter/4.java
@@ -1,1 +1,0 @@
-
