+++ src/main/java/use_case/create_task/CreateTaskInputData.java
@@ -1,4 +1,4 @@
-package use_case;
+package use_case.create_task;
 
 import entity.Task;
